Refactor task services: move TasksService and FeedbackService to shared directory, update import paths, and remove unused files. Moved task interface out of main service file and created dedicated model file.
