+++ my-task-manager/src/app/shared/task.model.ts
@@ -1,0 +1,7 @@
+export interface Task {
+  name: string;
+  description: string;
+  dueDate?: Date;
+  completed: boolean;
+  subtasks?: { subName: string; subCompleted: boolean }[];
+}
